Format message with leetcode usernames and discord user mentions below
Instead of formatting the results/predictions table with discord usernames,
keep the leetcode username (which is constant) and instead mention the users
in order of placement using their discord_user_id (also constant).

This removes the need to fetch the discord username of every user before
sending a message (and might make it easier for user to find themselves
on the table, as they expect their leetcode username anyway).