Modify get_top_users to return UserPrediction objects
Currently the bot has no need for this method because
it fetches predictions for individual users.

It is useful as a utility method when developing to
quickly obtain some prediction results.

May also be need it in the future if the number of users
registered with the discord bot becomes large.
Getting all predictions in batches would require fewer
requests from lccn.lbao.site than one per bot-registered user.