Add leetcode_ranking_utils.py
Utils to fetch the ranking (leaderboard) of each contest from the leetcode API.

This started as a script to settle a dispute over whether the recent increase
in contestants in the biweekly contests was due to more Indian participants
(potentially due to placements season - the answer seems to be, yes).

Half of the code is just playing around with different ways to do binary search
which is used just count how many users solved at least one problem. Good
candidate to delete in the future.

I may end up using this in the bot to fetch the full leaderboard right after
the contest and post an initial message with the ranking of the registered users
(this is useful because leetcode has no friends/user search functionality).

Once the predictions website has the data on predicted rating changes, the bot
can update the initial message to include those.