Add function to fetch multi-user rating predictions
lccn.lbao.site has an endpoint /contest-records/predicted-rating which allows
querying rating predictions for up to 26 users at a time.

Unfortunately, that endpoint only returns rating predictions, but not contest rank
or user country. It also requires specifying the user's data_region (US or CN) in
the query, which currently am not storing.

For now implemented a query to the endpoint in fetch_multiple_rating_predictions.

In the future, to obtain rating predictions, I may use this endpoint to reduce load on
lccn.lbao.site API. Two requirements if I want to maintain current functionality:
- Need to fetch user ranks from leetcode itself, which requires reading the whole leaderboard
- Need to fetch the user's data region (assuming it's constant) from leetcode when a user
  registers with the bot