Prevent two discord users from registering same leetcode username in a server
If two discord users register the same leetcode username in a server, they will appear twice
in the leaderboard with predictions.

While the same person may want to use different discord accounts in different servers, there
should never be a use case for two discord accounts to register the same leetcode username in
a server, so the /register_leetcode_user command now prevents that, informing the user of the reason.